+++ app/lib/cabinet-service.js
@@ -19,6 +19,7 @@ angular.module('lizard-nxt')
 
   timeseriesResource = Restangular.one('api/v1/timeseries/');
   events = Restangular.one('api/v1/events/');
+  var regions = Restangular.one('api/v1/regions/');
 
   // Wms getFeatureInfo goes through a proxy. Specify url as a param.
   wmsGetFeatureInfo = Restangular.one('proxy/');
@@ -69,7 +70,8 @@ angular.module('lizard-nxt')
       login: gettextCatalog.getString("Log in"),
       logout: gettextCatalog.getString("Log out"),
       profile: gettextCatalog.getString("Modify profile"),
-      version: gettextCatalog.getString("Double click for lizard version number"),
+      version: gettextCatalog.getString(
+        "Double click for lizard version number"),
       openMenu: gettextCatalog.getString("Open data menu"),
       closeMenu: gettextCatalog.getString("Close data menu"),
       transparency: gettextCatalog.getString("Adjust opacity"),
@@ -96,6 +98,7 @@ angular.module('lizard-nxt')
     geocode: geocodeResource,
     raster: rasterResource,
     timeseries: timeseriesResource,
-    wmsGetFeatureInfo: wmsGetFeatureInfo
+    wmsGetFeatureInfo: wmsGetFeatureInfo,
+    regions: regions
   };
 }]);