+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased (2.0.0) (XXXX-XX-XX)
 - Add region aggregation as a fourth aggregation tool. Draw regions and get
   raster aggregations when clicked.
 
+- Add doxx to build task.
+
 - Add angular-gettext for translations.
 
 - Add grunt tasks to extract and compile translations.